refactor(convert_qat.py): 更新模型导出格式并优化量化过程
- 使用新的魔数 'MLP4' 标识更新的模型格式
- 移除输入层的激活尺度，仅导出隐藏层的激活尺度
-修正 LUT 核心函数的参数类型
- 优化模型加载和预处理流程
- 移除冗余的内存计算和比较

+++ src/kernels.cpp
@@ -29,7 +29,7 @@ int hsum_i32_16(const __m512i a) {
 int32_t avx512_bit_slice_gemm_kernel(
     const uint8_t* input_packed_ptr,
     const uint8_t* weights_packed_ptr,
-    const int16_t* precomputed_lut_ptr,
+    const int32_t* precomputed_lut_ptr, // <-- CORRECTED: Was int16_t
     int k_dim
 ) {
     __m512i accumulated_sum_vec_32 = _mm512_setzero_si512();
@@ -267,4 +267,3 @@ void standard_linear_forward(
         }
     }
 }
-
